Day 12 - Star 2

+++ Day12.Tests/ArrangementCalculatorTest.cs
@@ -8,7 +8,7 @@ namespace Day12.Tests
         [TestCase("????.#...#... 4,1,1", ExpectedResult = 1)]
         [TestCase("????.######..#####. 1,6,5", ExpectedResult = 4)]
         [TestCase("?###???????? 3,2,1", ExpectedResult = 10)]
-        public int Test1(string data)
+        public long Test1(string data)
         {
             var x = new ArrangementCalculator(1);
             return x.CalculateArrangements(data);
@@ -20,7 +20,7 @@ namespace Day12.Tests
         [TestCase("????.#...#... 4,1,1", ExpectedResult = 16)]
         [TestCase("????.######..#####. 1,6,5", ExpectedResult = 2500)]
         [TestCase("?###???????? 3,2,1", ExpectedResult = 506250)]
-        public int Test5(string data)
+        public long Test5(string data)
         {
             var x = new ArrangementCalculator(5);
             return x.CalculateArrangements(data);
